feat: US-006 - Move DTOs to libs
