+++ src/auth/auth.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Post } from '@nestjs/common';
 import { AuthService } from '../../libs/auth/auth.service';
 import { Public } from '../../libs/auth/decorators/public.decorator';
-import { LoginResponseDto } from './dto/login-response.dto';
-import { LoginDto } from './dto/login.dto';
+import { LoginResponseDto } from '../../libs/auth/dto/login-response.dto';
+import { LoginDto } from '../../libs/auth/dto/login.dto';
 
 @Controller('api/auth')
 export class AuthController {
